requirement.txt
streamlit
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain_google_genai
langchain-ollama
langchain-community
